+++ test/widgets/game_start_dialog_test.dart
@@ -1,34 +1,33 @@
 import 'package:curling_scoreboard_flutter/l10n/app_localizations.dart';
+import 'package:curling_scoreboard_flutter/src/version.dart';
+import 'package:curling_scoreboard_flutter/widgets/game_start/game_start_dialog.dart';
 import 'package:flutter/material.dart';
-
-Widget wrapWithMaterialApp(Widget child) {
-  return MaterialApp(
-    localizationsDelegates: AppLocalizations.localizationsDelegates,
-    supportedLocales: AppLocalizations.supportedLocales,
-    home: Scaffold(body: child),
-  );
-}
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  /*testWidgets('GameStartDialog renders and allows selection', (
-    WidgetTester tester,
-  ) async {
-    await tester.pumpWidget(wrapWithMaterialApp(const GameStartDialog()));
+  testWidgets('GameStartDialog displays version number', (tester) async {
+    tester.view.physicalSize = const Size(3000, 1000);
+    tester.view.devicePixelRatio = 1.0;
 
-    // Check for title
-    expect(find.text('Game Setup'), findsOneWidget);
-    // Check for number of ends options
-    expect(find.text('2'), findsOneWidget);
-    expect(find.text('4'), findsOneWidget);
-    expect(find.text('6'), findsOneWidget);
-    expect(find.text('8'), findsOneWidget);
-    expect(find.text('10'), findsOneWidget);
-    // Check for players per team options
-    expect(find.text('0'), findsOneWidget);
-    expect(find.text('2'), findsWidgets); // 2 appears in both ends and players
-    expect(find.text('4'), findsWidgets);
-    // Check for hammer options
-    expect(find.text('Red'), findsOneWidget);
-    expect(find.text('Yellow '), findsOneWidget);
-  });*/
+    await tester.pumpWidget(
+      const MaterialApp(
+        localizationsDelegates: [
+          AppLocalizations.delegate,
+          GlobalMaterialLocalizations.delegate,
+          GlobalWidgetsLocalizations.delegate,
+          GlobalCupertinoLocalizations.delegate,
+        ],
+        supportedLocales: [
+          Locale('en', ''),
+        ],
+        home: GameStartDialog(),
+      ),
+    );
+
+    expect(find.text('v$packageVersion'), findsOneWidget);
+
+    addTearDown(tester.view.resetPhysicalSize);
+    addTearDown(tester.view.resetDevicePixelRatio);
+  });
 }